pyproject.toml
[tool.poetry]
name = "generative-ai-conference-demo"
version = "0.1.0"
description = ""
authors = ["Memorysaver <mfa1484@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.20.0"
watchdog = "^2.3.1"
streamlit-chat = "^0.0.2.2"
langchain = "^0.0.139"
python-dotenv = "^1.0.0"
chromadb = "^0.3.21"
tiktoken = "^0.3.3"


[virtualenvs]
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
